Improved the improvable

+++ src/main/java/ee/veebiprojekt/veebiprojekttest/entity/Joke.java
@@ -29,5 +29,5 @@ public class Joke {
     @Column(name = "rating")
     private double rating;
     @Column(name = "created_by")
-    private double createdBy;
+    private long createdBy;
 }
